all gui fixed
