+++ multinmrfit/ui.py
@@ -1,584 +1,697 @@
-# Import system libraries
-from tkinter.constants import N
+import tkinter
+import tkinter.messagebox
 import pkg_resources
 import random
-import json
-import sys
-import os 
-import logging
-import argparse
-import threading
-from pathlib import Path
-import warnings
-warnings.filterwarnings('ignore')
-from pathlib import Path 
-import customtkinter 
-#https://github.com/TomSchimansky/CustomTkinter
-# customtkinter.set_appearance_mode("System")
-# customtkinter.set_default_color_theme("dark-blue")
 
-# Import display libraries
+import sys
+import json
+from pathlib import Path 
 import tkinter as tk
 from tkinter import simpledialog, ttk, filedialog
 from PIL import Image, ImageTk
-
-# Import plot libraries
-import matplotlib
-import matplotlib.pyplot as plt
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
+import logging
 
 # Import math libraries
 import pandas as pd
 import numpy as np
 
-# Import our own libraries
-import multinmrfit.run as nmrr
-import multinmrfit.fitting as nff
+# Import plot libraries
+import matplotlib.pyplot as plt
+from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 
-matplotlib.use("TkAgg")
+import multinmrfit.io as nio
+import multinmrfit.run as nrun
+import multinmrfit.utils_nmrdata as nfu
 
-tk_wdw = None 
-logger = logging.getLogger()
+logger = logging.getLogger(__name__)
+# logger = logging.getLogger()
 
-class MyOptionMenu(tk.OptionMenu):
-    def __init__(self, master, status, *options):
-        self.var = tk.StringVar(master)
-        self.var.set(status)
-        tk.OptionMenu.__init__(self, master, self.var, *options)
-        self.config(font=('calibri',(10)),bg='white',width=12)
-        self['menu'].config(font=('calibri',(10)),bg='white')
+ 
+class App_Error:
+    def __init__(self, message, *args, **kwargs):
+        APP_NAME = "Error"
+        WIDTH = 500
+        HEIGHT = 100
 
-###################################
-# Peak Picking window
-###################################
+        super().__init__(*args, **kwargs)
+        master = tk.Tk()
+        self.master = master
+        master.title(APP_NAME)
+        master.geometry(str(WIDTH) + "x" + str(HEIGHT))
+        master.minsize(WIDTH, HEIGHT)
+        master.configure(bg='#2F4F4F')
+        master.protocol("WM_DELETE_WINDOW", self.on_closing)
+        master.toplevel = None
+        self.label = tk.Label(
+                            master, 
+                            text=message, 
+                            width= WIDTH,
+                            height=HEIGHT,
+                            bg='#2F4F4F',
+                            fg='white'
+                            )
+        self.label.pack()
+        # self.configure("TCombobox", fieldbackground= "orange", background= "white")
+        self.toplevel = None
+        if sys.platform == "darwin":
+            master.bind("<Command-q>", self.on_closing)
+            master.bind("<Command-w>", self.on_closing)
 
-def save_info_clustering(dic, Res):
-    Res.Peak_Intensity = dic['Peak Intensity']
-    Res.Peak_Position = dic['Peak Position']
-    Res.Options = [i.get() for i in dic["Options"]]
-    Res.Cluster = [i.get() for i in dic["Cluster ID"]]
-    Res.Selection = [True if i.get() != '' else False for i in dic["Cluster ID"]]
-    tk_wdw.destroy()
-    plt.close()
+    def on_closing(self, event=0):
+        self.master.destroy()
 
-def refresh_ui(refreshed_threshold, threshold_entry):
-    refreshed_threshold["th"] = threshold_entry.get()
-    tk_wdw.destroy()
-    plt.close()
+    def start(self):
+        self.master.mainloop()
 
-def Exit():
-    if tk.messagebox.askokcancel("quit",'Do you want to quit?'):
-        tk_wdw.destroy()
-        plt.close()
+class App:
+
+    def __init__(self, user_input, *args, **kwargs):
+        APP_NAME = "Multinmrfit Interface (v2.0)"
+        WIDTH = 1200
+        HEIGHT = 600
+        
+        MAIN_COLOR = "#5EA880"
+        OPTION_COLOR = "#001933"
+        BUTTON_COLOR = "#1c4587"
+
+        ENTRY_COLOR = "#3a8eba"
+        MAIN_HOVER = "#3a8eba"
+        FRAME_COLOR = '#708090'
+
+        super().__init__(*args, **kwargs)
+        master = tk.Tk()
+        self.master = master
+        master.title(APP_NAME)
+        master.geometry(str(WIDTH) + "x" + str(HEIGHT))
+        master.minsize(WIDTH, HEIGHT)
+        master.configure(bg='#2F4F4F')
+        master.protocol("WM_DELETE_WINDOW", self.on_closing)
+
+        # self.configure("TCombobox", fieldbackground= "orange", background= "white")
+        self.toplevel = None
+        if sys.platform == "darwin":
+            master.bind("<Command-q>", self.on_closing)
+            master.bind("<Command-w>", self.on_closing)
+        
+        # ============ create CTkFrames ============
+        self.frame_inputs = tk.Frame(master,
+                                    width=250,
+                                    height=HEIGHT-250,
+                                    bg=FRAME_COLOR
+                                    )
+        self.frame_inputs.place(relx=0.02, rely=0.1, anchor=tkinter.NW)
+
+        self.frame_analysis = tk.Frame(master,
+                                    width=250,
+                                    height=HEIGHT-250,
+                                    bg=FRAME_COLOR
+                                    )
+        self.frame_analysis.place(relx=0.25, rely=0.1, anchor=tkinter.NW)
+
+        self.frame_output = tk.Frame(master,
+                                    width=250,
+                                    height=HEIGHT-250,
+                                    bg=FRAME_COLOR
+                                    )
+        self.frame_output.place(relx=0.48, rely=0.1, anchor=tkinter.NW)
+
+        self.frame_options_Pseudo2D = tk.Frame(master,
+                                                width=250,
+                                                height=150,
+                                                bg=FRAME_COLOR
+                                                # fg_color=("white", "gray18")
+                                                )
+        self.frame_options_Pseudo2D.place(relx=0.71, rely=0.1, anchor=tkinter.NW)
+
+        self.frame_options_1DSeries = tk.Frame(master,
+                                                width=250,
+                                                height=150,
+                                                bg=FRAME_COLOR
+                                                # fg_color=("white", "gray18")
+                                                )
+        self.frame_options_1DSeries.place(relx=0.71, rely=0.4, anchor=tkinter.NW)
+
+         # ============ create CTkFrames ============
+        # Set bottom picture
+        path_image = pkg_resources.resource_filename('multinmrfit', 'data/')
+        img_network = Image.open(str(Path(path_image, 'network.png')))
+        self.img_network = ImageTk.PhotoImage(img_network.resize((800, 200)),Image.ANTIALIAS) 
+        self.img_network_label = tkinter.Label(master, image = self.img_network)
+        self.img_network_label.place(relx = 0.02, rely = 0.7)
+
+        # ============ create Labels and entries ============
+        title = ['Inputs','Analysis','Outputs']
+        for i in range(len(title)):
+            tk.Label(
+                master,
+                text=title[i].replace("_", " ").capitalize() ,
+                font=("Helvetica", 18, 'normal'),
+                bg=MAIN_HOVER,
+                fg='white',
+                width=10,
+                borderwidth=0,
+                justify=tk.CENTER
+            ).place(relx=0.02+i*0.23, rely=0.1, anchor=tkinter.W)
+
+
+        inputs=['data_path','data_folder','data_exp_no','data_proc_no']
+        for i in range(len(inputs)):
+            tk.Label(
+                self.frame_inputs,
+                width=10,
+                text=inputs[i].replace("_", " ").capitalize(),
+                font=("Helvetica", 14, 'normal'),
+                borderwidth=0,
+                fg='white',
+                bg=FRAME_COLOR,
+                justify=tk.CENTER
+            ).place(relx=0.5, rely=0.1+i*0.2, anchor="center")
+
+            self.input_var = tk.StringVar() 
+            self.input_entry = tk.Entry(
+                                    self.frame_inputs,
+                                    textvariable=self.input_var,
+                                    bg='white',
+                                    fg='black',
+                                    borderwidth=0,
+                                    )
+            self.input_entry.place(relx=0.1, rely=0.2+i*0.2, width=200, anchor=tkinter.W)
+            user_input[inputs[i]]  = self.input_var
+
+
+        analysis_info = ['analysis_type','reference_spectrum','spectral_limits','threshold']
+
+        style= ttk.Style()
+        style.theme_use('clam')
+        # style.configure("TCombobox", fieldbackground= "blue", background= "white")
+
+        for i in range(len(analysis_info)):
+            tk.Label(
+                self.frame_analysis,
+                text=analysis_info[i].replace("_", " ").capitalize(),
+                font=("Helvetica", 14, 'normal'),
+                width=20,
+                borderwidth=0,
+                fg='white',
+                bg=FRAME_COLOR,
+                #justify=tk.CENTER
+            ).place(relx=0.5, rely=0.1+i*0.2, anchor="center")
+
+            if analysis_info[i] == 'analysis_type':
+                self.analysis_info_var = tk.StringVar()
+                ttk.Combobox(
+                            self.frame_analysis,
+                            textvariable=self.analysis_info_var, 
+                            values=['1D','Pseudo2D','1D_Series'], 
+                            state="readonly",
+                            # bg='white',
+                            # fg='black',
+                ).place(relx=0.1, rely=0.2+i*0.2, width=200, anchor=tkinter.W)
+                user_input[analysis_info[i]]  = self.analysis_info_var
+
+            else:
+                self.analysis_info_var = tk.StringVar()
+                analysis_info_entry = tk.Entry(
+                                        self.frame_analysis,
+                                        textvariable=self.analysis_info_var,   
+                                        bg='white',
+                                        fg='black',
+                                        borderwidth=0,
+                                        )
+                analysis_info_entry.place(relx=0.1, rely=0.2+i*0.2, width=200, anchor=tkinter.W)
+                user_input[analysis_info[i]]  = self.analysis_info_var
+
+
+        outputs = ['output_path','output_folder','output_name']
+        for i in range(len(outputs)):
+            tk.Label(
+                self.frame_output,
+                text=outputs[i].replace("_", " ").capitalize() if not 'Options' in outputs[i] else outputs[i].replace("_", " ") ,
+                font=("Helvetica", 14, 'normal'),
+                width=20,
+                borderwidth=0,
+                fg='white',
+                bg=FRAME_COLOR,
+                #justify=tk.CENTER
+            ).place(relx=0.5, rely=0.1+i*0.2, anchor="center")
+
+            self.outputs_var = tk.StringVar()
+            outputs_entry = tk.Entry(
+                                self.frame_output,
+                                textvariable=self.outputs_var,
+                                bg='white',
+                                fg='black',
+                                borderwidth=0
+                                )
+            outputs_entry.place(relx=0.1, rely=0.2+i*0.2, width=200, anchor=tkinter.W)
+            user_input[outputs[i]]  = self.outputs_var
+
+
+        # ============ create Buttons ============
+        self.load_button = tk.Button(master,
+                                            text=" Load ",
+                                            highlightbackground =FRAME_COLOR,
+                                            fg='black',
+                                            borderwidth=0,
+                                            font=("Helvetica", 20, 'normal'),
+                                            command=lambda:nio.load_config_file(self.master,user_input))
+        self.load_button.place(relx=0.74, rely=0.70,width=80,height=30)
+
+        self.save_button = tk.Button(master,
+                                            text=" Save ",
+                                            highlightbackground =FRAME_COLOR,
+                                            fg='black',
+                                            borderwidth=0,
+                                            font=("Helvetica", 20, 'normal'),
+                                            command=lambda:self.save_config_file({k: v.get() for k, v in user_input.items()}))
+        self.save_button.place(relx=0.81, rely=0.70,width=80,height=30)
+
+        self.run_button = tk.Button(master,
+                                            text=" Run ",
+                                            highlightbackground =FRAME_COLOR,
+                                            fg='black',
+                                            borderwidth=0,
+                                            font=("Helvetica", 20, 'normal'),
+                                            command=lambda:self.App_Run(user_input))
+        self.run_button.place(relx=0.74, rely=0.80,width=80,height=30)
+
+        self.close_button = tk.Button(master,
+                                            text=" Close ",
+                                            highlightbackground =FRAME_COLOR,
+                                            fg='black',
+                                            borderwidth=0,
+                                            font=("Helvetica", 20, 'normal'),
+                                            command=lambda:self.on_closing())
+        self.close_button.place(relx=0.81, rely=0.80,width=80,height=30)
+
+        # # ============ Options ============
+        # option_list = []
+        tk.Label(
+                master,
+                text='Options: Pseudo2D',
+                width=20,
+                font=("Helvetica", 18, 'normal'),
+                bg=MAIN_HOVER,
+                fg='white',
+                borderwidth=0,
+                justify=tk.CENTER
+        ).place(relx=0.02+3*0.23, rely=0.1, anchor=tkinter.W)
+
+        tk.Label(
+                master,
+                text='Options: 1DSeries',
+                width=20,
+                font=("Helvetica", 18, 'normal'),
+                bg=MAIN_HOVER,
+                fg='white',
+                borderwidth=0,
+                justify=tk.CENTER
+        ).place(relx=0.02+3*0.23, rely=0.4, anchor=tkinter.W)
+
+
+        self.TimeSeries = tk.IntVar()
+        self.check_box_1 = tk.Checkbutton(
+                                self.frame_options_1DSeries,
+                                text="Time Series",
+                                font=("Helvetica", 14, 'normal'),
+                                variable=self.TimeSeries,
+                                highlightthickness=0,
+                                bd=0,
+                                bg=FRAME_COLOR
+                                )
+        self.check_box_1.place(relx=0.25, rely=0.2, anchor=tkinter.W)
+        user_input['time_series'] = self.TimeSeries
+
+        tk.Label(
+            self.frame_options_Pseudo2D,
+            text='raw ids',
+                width=20,
+                font=("Helvetica", 14, 'normal'),
+                borderwidth=0,
+                fg='white',
+                bg=FRAME_COLOR,
+                justify=tk.CENTER
+             ).place(relx=0.5, rely=0.25, anchor="center")
+       
+        self.input_raws = tk.StringVar()
+        input_entry = tk.Entry(
+                            self.frame_options_Pseudo2D,
+                            textvariable=self.input_raws,
+                            bg='white',
+                            fg='black',
+                            borderwidth=0
+                            )
+        input_entry.place(relx=0.1, rely=0.47, width=200, anchor=tkinter.W)
+        user_input['data_row_no'] = self.input_raws
+
+    def App_Run(self, user_input):
+        user_input = nio.check_input_file({k: v.get() for k, v in user_input.items()},self.master)
+        nrun.run_analysis(user_input,self)
+        self.on_closing()
+
+    def ask_filename(self,config_path, event=0):
+        wdw = tk.Tk()
+        wdw.withdraw()
+        # the input dialog
+        file_name = simpledialog.askstring(
+            title="Config File Name",
+            prompt="Config Input File Name:",
+            initialvalue="Inputs_Spec_Fitting")
+        if file_name is None:
+            wdw.destroy()
+            config_file = None
+        else:
+            config_file = Path(config_path,str(file_name)+'.json')
+            wdw.destroy()
+        return config_file
+
+    def save_config_file(self,user_input): 
+        config_path = Path(user_input['output_path'], user_input['output_folder'])
+        file_name = self.ask_filename(config_path)
+
+        if not 'rows_pseudo2D' in user_input:
+            pass
+        else:
+            if not user_input['rows_pseudo2D']:
+                del user_input['rows_pse    udo2D']
+        if file_name :
+            f = open(file_name, "a")
+            f.seek(0)
+            f.truncate()
+            f.write(json.dumps(user_input, indent=4))
+            f.close()  
+
+    def on_closing(self, event=0):
+        self.master.destroy()
         exit()
 
-def user_clustering_gui(peak_picking_threshold, peak_picking_data, figure, colors, Res):
-    global tk_wdw
-    tk_wdw = tk.Tk()
-    tk_wdw.protocol("WM_DELETE_WINDOW", Exit)
-    tk_wdw.title('Peak Picking Visualisation and Validation')
+    def start(self):
+        self.master.mainloop()
 
-    graph = FigureCanvasTkAgg(figure, master=tk_wdw)
-    canvas = graph.get_tk_widget()
-    canvas.grid(row=0, column=0,columnspan = 8,rowspan=1)
+class App_Clustering:
 
-    list_of_options=['Roof'] # options
     
-    user_input = {
-        'Peak Position'    :   [],
-        'Peak Intensity'   :   [],
-        'Cluster ID'       :   [],
-        'Options'          :   []
-    }
-
-    npeaks = len(peak_picking_data)
-    if not npeaks:
-        tk.Label(
-            tk_wdw, 
-            text='No peak found, please lower the threshold',
-            font=("Helvetica", 14, 'bold'),
-            fg='#f00'
-        ).grid(column=0, row=3)
-        tk.Entry(tk_wdw,justify = "center")
-    else:
-        c = 0 
-        for label in user_input.keys():
-            tk.Label(
-                tk_wdw, 
-                text=label, 
-                font=("Helvetica", 14, 'bold')
-            ).grid(column=c+1, row=2)
-            c +=1
-
-        for i in range(npeaks):
-            tk.Label(tk_wdw, text="Peak "+str(i+1),fg=colors[i]).grid(column=0, row=i+3)
-            tk.Entry(tk_wdw,justify = "center")
+    def __init__(self, x_spec, y_spec, peak_picking_threshold, clustering_table, *args, **kwargs):
+        self.APP_NAME = "Peak Picking Visualisation and Clustering"
+        self.WIDTH = 1200
+        self.HEIGHT = 600
         
-            # Clustering
-            cluster_entry = tk.Entry(tk_wdw,justify = "center")
-            user_input['Cluster ID'].append(cluster_entry)
-            cluster_entry.grid(row=i+3,column=3)
+        MAIN_COLOR = "#5EA880"
+        self.ENTRY_COLOR = "#3c78d8"
+        OPTION_COLOR = "#001933"
+        BUTTON_COLOR = "#1c4587"
+        self.MAIN_HOVER = "#3a8eba"
 
-            options_entry = ttk.Combobox(tk_wdw, values=list_of_options)
-            user_input['Options'].append(options_entry)
-            options_entry.grid(row=i+3,column=4)
-            for col in peak_picking_data.columns:
-                en_c = tk.Entry(tk_wdw,justify = "center")
-                data = peak_picking_data.iloc[i].loc[col]
-                if col == 'Peak_Position':
-                    user_input['Peak Position'].append(data)
-                    cc = 0
-                if col == 'Peak_Intensity':
-                    user_input['Peak Intensity'].append(data)
-                    cc = 1
-                en_c.insert(0, round(data,3))
-                en_c.grid(column=cc+1,row=i+3)
+        self.FRAME_COLOR = '#708090'
+
+        super().__init__(*args, **kwargs)
+        master = tk.Tk()
+        self.master = master
     
-    tk.Label(
-        tk_wdw, 
-        text="Threshold", 
-        font=("Helvetica", 14, 'bold'), fg='#f00'
-    ).grid(column=0, row=1)
+        master.title(self.APP_NAME)
+        master.geometry(str(self.WIDTH) + "x" + str(self.HEIGHT))
+        master.minsize(self.WIDTH, self.HEIGHT)
+        master.configure(bg='#2F4F4F')
 
-    threshold_entry = tk.Entry(tk_wdw, readonlybackground="white")
-    threshold_entry.insert(0, peak_picking_threshold)
-    threshold_entry.grid(column=1, row=1)
-    refreshed_threshold = { "th" : None }
+        master.protocol("WM_DELETE_WINDOW", self.on_closing)
+        self.widget = None
+        
+        self.df_var = tk.StringVar()
 
-    tk.Button(
-        tk_wdw, 
-        text="Refresh & Close", 
-        fg = "orange", 
-        font=("Helvetica", 20),        
-        command=lambda: refresh_ui(refreshed_threshold, threshold_entry)
-    ).grid(column = 0, row = npeaks+3 if npeaks else 4)  
-    
-    tk.Button(
-        tk_wdw, 
-        text="Save & Close", 
-        fg = "blue", 
-        font=("Helvetica", 20),        
-        command=lambda: save_info_clustering(user_input, Res)
-    ).grid(column = 1, row = npeaks+3 if npeaks else 4) 
+        # self.configure("TCombobox", fieldbackground= "orange", background= "white")
+        self.toplevel = None
+        if sys.platform == "darwin":
+            master.bind("<Command-q>", self.on_closing)
+            master.bind("<Command-w>", self.on_closing)
+        # # ============ create CTkFrames ============
+        self.frame_graph = tk.Frame(master,
+                                    width=500,
+                                    height=self.HEIGHT-150,
+                                    bg=self.FRAME_COLOR
+                                    )
+        self.frame_graph.place(relx=0.02, rely=0.03, anchor=tkinter.NW)
 
-    tk.Button(
-        tk_wdw, 
-        text="Exit", 
-        fg = "black", 
-        font=("Helvetica", 20),        
-        command=lambda: Exit()
-    ).grid(column = 2, row = npeaks+3 if npeaks else 4) 
+        self.frame_threshold = tk.Frame(master,
+                                    width=500,
+                                    height=self.HEIGHT-520,
+                                    bg=self.FRAME_COLOR
+                                    )
+        self.frame_threshold.place(relx=0.02, rely=0.82, anchor=tkinter.NW)
 
-    tk_wdw.mainloop()
-    if refreshed_threshold["th"]:
-        return float(refreshed_threshold["th"])
-    else:
-        return None
 
-def filter_multiple_clusters(Res):
-    for i,j in enumerate(Res.Cluster):
-        cluster_num = j.split(',')
-        if len(cluster_num) > 1:
-            for k in cluster_num:
-                new_pk = {'Peak_Position': Res.iloc[i].ppm_H_AXIS,'Peak_Intensity': Res.iloc[i].Peak_Amp,'Selection': Res.iloc[i].Selection,'Cluster': k}
-                Res = Res.append(new_pk, ignore_index = True)
-            Res = Res.drop(Res.index[i])
-    return Res
+        self.frame_peak_Table = tk.Frame(master,
+                                    width=640,
+                                    height=self.HEIGHT-50,
+                                    bg=self.FRAME_COLOR
+                                    )
+        self.frame_peak_Table.place(relx=0.47, rely=0.03, anchor=tkinter.NW)
 
-def plot_picking_data(x_spec, y_spec, peak_picking_threshold, peak_picking_data):
-    n_peak = len(peak_picking_data)
+        # ============ Figure ============
+        peak_picking_data = self.peak_picking(x_spec, y_spec, peak_picking_threshold)
+        colors = self.create_plot(x_spec, y_spec, peak_picking_threshold,peak_picking_data)
+        clustering_information = self.create_table(peak_picking_data,colors) 
 
-    # Create a list of colors
-    colors = []
-    for i in range(n_peak):
-        colors.append('#%06X' % random.randint(0, 0xFFFFFF))
+        # ============ Labels ============
+        tk.Label(
+            self.frame_threshold,
+            text='Threshold',
+            width=10,
+            borderwidth=0,
+            fg='white',
+            bg=self.MAIN_HOVER,
+            font=("Helvetica", 18, 'normal'),
+            justify=tk.CENTER
+        ).place(relx=0.02, rely=0.2, anchor=tkinter.W)
 
-    fig = plt.figure()
-    plt.plot(x_spec, y_spec, '-',color='teal')
-    for i in range(n_peak):
-        plt.plot(
-            peak_picking_data.Peak_Position.iloc[i],
-            peak_picking_data.Peak_Intensity.iloc[i],
-            c=colors[i],
-            ls='none',
-            marker='o'
-            )
-    plt.axhline(peak_picking_threshold,c='r')
-    plt.gca().invert_xaxis()
-    plt.xlabel(r'$^1H$ $(ppm)$')
-    return fig, colors 
 
-def run_user_clustering(figures, colors_plot, peak_picking_threshold, peak_picking_data):
+        # ============ Entry ============
+        # self.threshold_var = tk.StringVar()
+        self.threshold_entry = tk.Entry(
+                                    self.frame_threshold,
+                                    justify = "center",
+                                    width=12,
+                                    bg='white',
+                                    fg='black',
+                                    borderwidth=0,
+                                    )
+        self.threshold_entry.insert(0, peak_picking_threshold)                                  
+        self.threshold_entry.place(relx=0.4, rely=0.2, width=200, anchor=tkinter.W)
 
-    clustering_results = pd.DataFrame(columns=['Peak_Position','Peak_Intensity','Selection','Cluster','Options'])
+        # ============ Buttons ============
+        self.refresh_th = tk.Button(master,
+                                    text=" Refresh & Close ",
+                                    highlightbackground = self.FRAME_COLOR,
+                                    fg='black',
+                                    borderwidth=0,
+                                    font=("Helvetica", 20, 'normal'),
+                                    command=lambda:self.refresh_ui(x_spec, y_spec, self.threshold_entry.get(),colors)
+                                    )
+          
+        self.refresh_th.place(relx=0.54, rely=0.85,width=180,height=50)
 
-    refreshed_threshold = user_clustering_gui(
-        peak_picking_threshold,
-        peak_picking_data, 
-        figures,
-        colors_plot,
-        clustering_results
-    )
+        self.run = tk.Button(master,
+                            text=" Run Fitting ",
+                            highlightbackground = self.FRAME_COLOR,
+                            fg='black',
+                            borderwidth=0,
+                            font=("Helvetica", 20, 'normal'),
+                            command=lambda:self.save_info_clustering(clustering_information, clustering_table)
+                            )
+        self.run.place(relx=0.71, rely=0.85,width=180,height=50)
 
-    clustering_results = filter_multiple_clusters(clustering_results)
-    return refreshed_threshold, clustering_results
+        self.close_button = tk.Button(master,
+                            text=" Close ",
+                            highlightbackground = self.FRAME_COLOR,
+                            fg='black',
+                            borderwidth=0,
+                            font=("Helvetica", 20, 'normal'),
+                            command=lambda:self.on_closing())
+        self.close_button.place(relx=0.88, rely=0.85,width=80,height=50)
 
-###################################
-# Final Plots
-###################################
-    
-###################################
-# Error interface
-###################################
-
-def error_interface(message, critical_error=False):
-    if tk_wdw:
-        error_window = tk.Tk()
-        error_window.title("Error")
-        error_window.configure(bg='#FFFFFF')
-
-        label = tk.Label(error_window, text=message, font=("Helvetica",18),bg='#FFFFFF',borderwidth=20)
-        label.pack()
-
-        close_button = tk.Button(
-            error_window,
-            text="Close ",
-            # fg='#FFFFFF',
-            font=("Helvetica", 14),
-            # highlightbackground = "#0000FF",
-            command=lambda:error_window.destroy()
-
+    def peak_picking(self, x_spec_ref, y_spec_ref, threshold):
+        peak_picking = nfu.Peak_Picking_1D(
+            x_data          =   x_spec_ref, 
+            y_data          =   y_spec_ref, 
+            threshold       =   threshold,
         )
-        close_button.pack()
-    elif logger:
-        logger.error(message)
-    
-    if critical_error:
-        exit(1)
+        peak_picking = nfu.sort_peak_picking_data(peak_picking, 15)        
+        return peak_picking
 
-###################################
-# Loading Data interface
-###################################
-def create_experiments_list(user_input):
-    experiment_list = []
-    for i in user_input.split(','):
-        if "-" in i:
-            spectra = i.split('-')
-            experiment_list += range(int(spectra[0]), int(spectra[1])+1)
-        else:
-            experiment_list.append(int(i))
-    return experiment_list
+    def create_table(self,peak_picking_data,colors):
+        n_peak = len(peak_picking_data)
 
-# def check_path(path):
-#     """
-#     check if 'path' exists, and create it if it doesn't exist
-#     """
-#     sub_path = os.path.dirname(path)
-#     if not os.path.exists(sub_path):
-#         check_path(sub_path)
-#     if not os.path.exists(path):
-#         os.mkdir(path)
-
-def launch_analysis(user_input):
-    is_gui = (tk_wdw != None)
-    is_not_gui = (tk_wdw == None)
-
-    try:
-        output_dir = Path(user_input.get('output_path'),user_input.get('output_folder'))
-        # if not output_dir:
-        #     return error_interface("Argument : 'output_folder' is missing", critical_error=is_not_gui)
-
-        output_dir.mkdir(parents=True,exist_ok=True)
-
-        # create logger (should be root to catch builder and simulator loggers)
-        # logger = logging.getLogger()
-        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s', "%Y-%m-%d %H:%M:%S")
-        # sends logging output to 'process_log.txt' file
-        file_handler = logging.FileHandler(str(Path(output_dir, "process_log.txt")), mode='w+')
-        file_handler.setFormatter(formatter)
-        # sends logging output to sys.stderr
-        strm_handler = logging.StreamHandler()
-        strm_handler.setFormatter(formatter)
-        # add handlers to logger
-        logger.addHandler(strm_handler)
-        logger.addHandler(file_handler)
-        logger.setLevel(logging.INFO)
-
-        if not Path(user_input.get('data_path')).exists():
-            return error_interface("Argument : 'data_path' does not exist", critical_error=is_not_gui)
-
-        if not Path(user_input.get('data_path'),user_input.get('data_folder')).exists():
-            return error_interface("Argument : 'data_folder' does not exist", critical_error=is_not_gui)
-
-        if float(user_input.get('threshold', 1)) <= float(0):
-            return error_interface("Argument : 'threshold' is too low (should be > 0)", critical_error=is_not_gui)
-
-        if user_input.get('analysis_type') not in ['Pseudo2D', '1D', '1D_Series']:
-            return error_interface("Argument : 'analysis_type' expected as 'Pseudo2D','1D' or '1D_Series", critical_error=is_not_gui)
-
-        if not user_input.get('spectral_limits'):
-            return error_interface("Argument : 'spectral_limits' is missing" , critical_error=is_not_gui)
-
-        spec_lim = [float(i) for i in user_input.get('spectral_limits').split(',')]
-
-        if len(spec_lim)%2 != 0 and len(spec_lim) != 0:
-            return error_interface("Argument : 'spectral_limits' is incomplete", critical_error=is_not_gui)
-
-        exp_list = create_experiments_list(user_input.get('data_exp_no'))
-        for exp in exp_list:
-            if not Path(user_input.get('data_path'),user_input.get('data_folder'),str(exp)).exists():
-                return error_interface(f"Argument : experiment <{exp}> does not exist", critical_error=is_not_gui)
-            else:
-                if not Path(user_input.get('data_path'),user_input.get('data_folder'),str(exp),'pdata',user_input.get('data_proc_no')).exists():
-                    return error_interface(f"Argument : experiment/procno <{exp}/{user_input.get('data_proc_no')}> does not exist", critical_error=is_not_gui)
-        if user_input.get('analysis_type') != 'Pseudo2D' and int(user_input.get('reference_spectrum')) not in exp_list:
-            return error_interface(f"Argument : reference_spectrum <{user_input.get('reference_spectrum')}> not found in experiment list", critical_error=is_not_gui)
-
-        row_list = []
-        if user_input.get('data_row_no'):
-            row_list = create_experiments_list(user_input.get('data_row_no'))
-            if int(user_input.get('reference_spectrum')) not in row_list :
-                return error_interface(f"Argument : reference_spectrum <{user_input.get('reference_spectrum')}> not found in row list", critical_error=is_not_gui)
-
-        config = {
-            'data_path'             :   user_input.get('data_path'),
-            'data_folder'           :   user_input.get('data_folder'),
-            'data_exp_no'           :   exp_list,
-            'data_proc_no'          :   int(user_input.get('data_proc_no')),
-            'data_row_no'           :   row_list,
-            'reference_spectrum'    :   user_input.get('reference_spectrum'),
-            'analysis_type'         :   user_input.get('analysis_type'),
-            'spectral_limits'       :   spec_lim,
-            'threshold'             :   float(user_input.get('threshold', 0)),
-            'output_path'           :   user_input.get('output_path'),
-            'output_folder'         :   user_input.get('output_folder'),
-            'output_name'           :   user_input.get('output_name'),
-
+        clustering_information = {
+            'Peak Position'    :   [],
+            'Peak Intensity'   :   [],
+            'Cluster ID'       :   [],
+            'Options'          :   []
         }
 
-    except Exception as e:
-        return error_interface(e, critical_error=is_not_gui)
-    for key, conf in config.items():        
-        if conf is None:
-            return error_interface(f"Argument : '{key}' is missing", critical_error=is_not_gui)
-    if is_gui:
-        tk_wdw.destroy()
-    logger.info(json.dumps(config,indent=4))
-    nmrr.run_analysis(config, gui = is_gui)
+        options = ['Roof'] # options
 
-def on_closing():
-    tk_wdw.destroy()
-    exit(0)
-
-def ask_filename(config_path):
-    wdw = tk.Tk()
-    wdw.withdraw()
-    # the input dialog
-    file_name = simpledialog.askstring(
-        title="Config File Name",
-        prompt="Config Input File Name:",
-        initialvalue="Inputs_Spec_Fitting")
-    if file_name is None:
-        wdw.destroy()
-    else:
-        config_file = Path(config_path,str(file_name)+'.json')
-        wdw.destroy()
-    return config_file
-
-def save_config_file(user_input):
-
-    config_path = Path(user_input['output_path'], user_input['output_folder'])
-    config_path.mkdir(parents=True,exist_ok=True)
-    file_name = ask_filename(config_path)
-    f = open(file_name, "a")
-    f.seek(0)
-    f.truncate()
-    f.write(json.dumps(user_input, indent=4))
-    f.close()    
-
-def create_entry(label, x, y, width=210):
-    # Label
-    create_label(label, x, y)
-
-    # Entry
-    if label == 'analysis_type':
-        analysis_var = tk.StringVar()
-        ttk.Combobox(
-            textvariable=analysis_var, 
-            values=['1D','Pseudo2D','1D_Series'], 
-            state="readonly"
-        ).place(x=x, y=y+30, width=width)
-        return analysis_var
-
-    imput_var = tk.StringVar()
-    input_entry = customtkinter.CTkEntry(textvariable=imput_var,corner_radius=8)
-    input_entry.place(x=x, y=y+30, width=width)
-
-    return imput_var
-
-def create_label(label, x, y):
-    customtkinter.CTkLabel(
-        tk_wdw,
-        text=label.replace("_", " ").capitalize(),
-        corner_radius=8,
-        borderwidth=0,
-        justify=tk.CENTER
-    ).place(x=x, y=y)
-
-def load_config_file(user_input=None, config_file_path=None):
-    if tk_wdw:
-        config_file_path = filedialog.askopenfilename()    
-        if not config_file_path:
-            return None
-
-    try:
-        f = open(str(Path(config_file_path)),"r")    
-    except FileNotFoundError:
-        error_interface('Config file not found')
-        return None
-
-    try:
-        config = json.loads(f.read())
-    except json.decoder.JSONDecodeError as e:
-        error_interface('Json config file must be reformated')
-        return None
-    except UnicodeDecodeError as e:
-        error_interface('Wrong config file type (not a json file, see example)')
-        return None
-
-    if tk_wdw:
-        for label in user_input.keys():
-            user_input[label].set(config.get(label, ''))
-        return user_input
-    else:
-        return config
-
-def start_gui():
-
-    user_input = {
-        'data_path':            None,
-        'data_folder':          None,
-        'data_exp_no':          None,
-        'data_proc_no':         None,
-        'analysis_type':        None,
-        'reference_spectrum':   None,    
-        'spectral_limits':      None,
-        'threshold':            None,
-        'output_path':          None,
-        'output_folder':        None,    
-        'output_name':          None,
-        'rows_pseudo2D':         None,
-
-    }
-    global tk_wdw
-    tk_wdw = customtkinter.CTk()
-    tk_wdw.title("Multinmrfit Interface")
-    tk_wdw.geometry("950x650")
-    #tk_wdw.configure(bg='#FFFFFF')
-
-    path_image = pkg_resources.resource_filename('multinmrfit', 'data/')
-    # Set bottom picture
-    img_network = Image.open(str(Path(path_image, 'network.png')))
-    img_network_ = ImageTk.PhotoImage(img_network.resize((950, 160))) 
-    img_network_label = tk.Label(tk_wdw, image = img_network_)
-    img_network_label.place(x = 00, y = 490)
-
-    # Import Logo
-    img_logo = Image.open(str(Path(path_image, 'logo_small.png')))
-    img_logo_ = ImageTk.PhotoImage(img_logo.resize((300, 100))) 
-    img_logo_logo = tk.Label(tk_wdw,image=img_logo_)
-    img_logo_logo.place(x = 300, y = 0)
-
-    title = ['Inputs','Analysis','Outputs']
-    i = 0
-    n_row = 4 
-    for label in user_input.keys():
-        x = 10 + int(i / n_row) * 240
-        y = 160 + int(i % n_row) * 70
-        if int(i/n_row) != 3:
-            if int(i%n_row) == 0:
-                create_label(title[int(i/n_row)], x + 50, 120)
-        user_input[label] = create_entry(label, x, y)
-        if label is 'output_name':
-            i += 2
+        if not n_peak:
+            th_label = tk.Label(
+                self.frame_threshold, 
+                text='No peak found, please lower the threshold',
+                font=("Helvetica", 18, 'bold'),
+                borderwidth=0,
+                fg='white',
+                bg=self.FRAME_COLOR,
+            )
+            th_label.place(relx=0.25, rely=0.8, anchor=tkinter.W)   
         else:
-            i += 1
+            if len(self.frame_threshold.winfo_children()) >2 :
+                self.frame_threshold.winfo_children()[2].destroy()
 
-    #CheckBox
-    check_box_1 = customtkinter.CTkCheckBox(
-        tk_wdw,
-        text="Options")
-    check_box_1.place(
-        x=850, 
-        y=120, 
-        anchor=tk.CENTER
-        )
-
-    ## ----- General Buttons ----- ##
-    LoadButton = customtkinter.CTkButton(
-        tk_wdw,
-        text=" Load ",
-        #fg='#FFFFFF',
-        #font=("Helvetica", 20),
-        corner_radius=10,
-#        highlightbackground = "#8B0000",
-        command=lambda:load_config_file(user_input)
-        )
-    LoadButton.place(x=20, y=440,width=80,height=30)
-    
-    SaveButton = customtkinter.CTkButton(
-        tk_wdw,
-        text=" Save ",
-        #fg='#FFFFFF',
-        #font=("Helvetica", 20),
-        corner_radius=10,
-#        highlightbackground = "#0000FF",
-        command=lambda:save_config_file({k: v.get() for k, v in user_input.items()})
-        )
-    SaveButton.place(x=200, y=440,width=80,height=30)
-
-    RunButton = customtkinter.CTkButton(
-        tk_wdw,
-        text=" Run ",
-        #fg='#FFFFFF',
-        #font=("Helvetica", 20),
-        corner_radius=10,
-        #highlightbackground = "#8B0000",
-        command=lambda:launch_analysis({k: v.get() for k, v in user_input.items()})
-
-    )
-    RunButton.place(x=380, y=440,width=80,height=30)
-    
-    CloseButton = customtkinter.CTkButton(
-        tk_wdw,
-        text=" Close ",
-        #fg='#FFFFFF',
-        #font=("Helvetica", 20),
-        corner_radius=10,
-        #highlightbackground = "#0000FF",
-        command=lambda:on_closing()
-        )
-    CloseButton.place(x=560, y=440,width=80,height=30)
-    
-    tk_wdw.mainloop()
-
-###################################
-# Progress bars
-###################################
-
-def progress_bar_exit(root):
-    root.destroy()
-
-# def progress_label(progress_bar):
-#     return f"Current Progress: {progress_bar['value']}%"
+            tk.Label(
+                self.frame_peak_Table,
+                text='Clutering',
+                width=10,
+                borderwidth=0,
+                fg='white',
+                bg=self.MAIN_HOVER,
+                font=("Helvetica", 18, 'normal'),
+                justify=tk.CENTER
+            ).grid(column=0, row=0)
 
 
+            c = 0 
+            for label in clustering_information.keys():
+                tk.Label(
+                    self.frame_peak_Table, 
+                    text=label, 
+                    borderwidth=0,
+                    fg='white',
+                    bg=self.FRAME_COLOR,
+                    width=10
+                    #font=("Helvetica", 14, 'bold')
+                ).grid(column=c+1, row=2,padx=2)
+                c +=1
+                
+            for i in range(n_peak):
+                peak_label = tk.Label(
+                    self.frame_peak_Table, 
+                    text="Peak "+str(i+1),
+                    borderwidth=0,
+                    # fg='white',
+                    bg=self.FRAME_COLOR,
+                    fg=colors[i]
+                    )
+                peak_label.grid(column=0, row=i+3,pady=5)
+
+                # Clustering
+                self.cluster_entry = tk.Entry(
+                                            self.frame_peak_Table,
+                                            justify = "center",
+                                            width=12,
+                                            bg='white',
+                                            fg='black',
+                                            borderwidth=0,
+                                            )
+                clustering_information['Cluster ID'].append(self.cluster_entry)
+                self.cluster_entry.grid(row=i+3,column=3)
+
+                # Options
+                self.options_entry = ttk.Combobox(
+                                            self.frame_peak_Table, 
+                                            values=options,
+                                            width=12
+                                            )
+                clustering_information['Options'].append(self.options_entry)
+                self.options_entry.grid(row=i+3,column=4)
+
+
+                # Positions and Intensities
+                for col in peak_picking_data.columns:
+                    self.entry_c = tk.Entry(
+                                            self.frame_peak_Table,
+                                            justify = "center",
+                                            width=12,
+                                            bg='white',
+                                            fg='black',
+                                            borderwidth=0,
+                                            )
+                    data = peak_picking_data.iloc[i].loc[col]
+                    if col == 'Peak_Position':
+                        clustering_information['Peak Position'].append(data)
+                        cc = 0
+                    if col == 'Peak_Intensity':
+                        clustering_information['Peak Intensity'].append(data)
+                        cc = 1
+                    self.entry_c.insert(0, round(data,3))
+                    self.entry_c.grid(column=cc+1,row=i+3)
+        
+
+        return clustering_information
+
+    def save_info_clustering(self, clustering_information, clustering_table):
+
+        clustering_table.Peak_Intensity = clustering_information['Peak Intensity']
+        clustering_table.Peak_Position = clustering_information['Peak Position']
+        clustering_table.Options = [i.get() for i in clustering_information["Options"]]
+        clustering_table.Cluster = [i.get() for i in clustering_information["Cluster ID"]]
+        clustering_table.Selection = [True if i.get() != '' else False for i in clustering_information["Cluster ID"]]
+        self.master.destroy()
+
+    def create_plot(self, x_spec, y_spec, threshold,peak_picking_data):
+        
+        # remove old widgets
+        if self.widget:
+            self.widget.destroy()
+
+        n_peak = len(peak_picking_data)
+
+        # Create a list of colors
+        colors = []
+        for i in range(n_peak):
+            colors.append('#%06X' % random.randint(0, 0xFFFFFF))
+
+        fig = plt.Figure(figsize=(6,5), dpi=100)
+        ax1 = fig.add_subplot(111)
+        ax1.plot(x_spec, y_spec, '-',color='teal')
+        for i in range(n_peak):
+            ax1.plot(
+                peak_picking_data.Peak_Position.iloc[i],
+                peak_picking_data.Peak_Intensity.iloc[i],
+                c=colors[i],
+                ls='none',
+                marker='o'
+                )
+        ax1.axhline(float(threshold),c='r')
+        ax1.invert_xaxis()
+        ax1.set_xlabel(r'$^1H$ $(ppm)$')
+
+
+        self.graph = FigureCanvasTkAgg(fig, self.frame_graph)
+        self.graph_canvas = self.graph.get_tk_widget()
+        self.graph_canvas.place(relx=0.0,rely=0.0,width=500,height=self.HEIGHT-150)
+
+        return colors
+
+    def refresh_ui(self, x_spec, y_spec, threshold, colors):
+        peak_picking = self.peak_picking(x_spec, y_spec, float(threshold))
+        colors = self.create_plot(x_spec, y_spec, threshold, peak_picking)
+        self.clear_frame()
+        self.create_table(peak_picking,colors)
+
+    def clear_frame(self):
+        for widgets in self.frame_peak_Table.winfo_children():
+            widgets.destroy()
+        
+    def on_closing(self, event=0):
+        self.master.destroy()
+        exit()
+
+    def start(self):
+        self.master.mainloop()
+
+##########
 def init_progress_bar_windows(len_progresses, title, progress_bar_label):
     root = tk.Tk()
     root_height= len(len_progresses)*120
     root.geometry(f'300x{root_height}')
     root.title(title)
-
+    root.configure(bg='#2F4F4F')
     progress_bars = []
     for len_progress in len_progresses:
         pg_bar = ttk.Progressbar(
@@ -590,7 +703,16 @@ def init_progress_bar_windows(len_progresses, title, progress_bar_label):
         )
 
         # value_label = ttk.Label(root, text=update_progress_label())
-        value_label = ttk.Label(root, text=progress_bar_label[len(progress_bars)])
+        value_label = tk.Label(
+                        root,
+                        text=progress_bar_label[len(progress_bars)],
+                        width=10,
+                        borderwidth=0,
+                        fg='white',
+                        bg='#2F4F4F',
+                        font=("Helvetica", 18, 'normal'),
+                        justify=tk.CENTER
+)
         value_label.grid(column=0, row=len(len_progresses)*len(progress_bars), columnspan=2)
 
         pg_bar.grid(column=0, row=len(len_progresses)*len(progress_bars)+1, columnspan=2, padx=10, pady=20)
@@ -601,57 +723,14 @@ def init_progress_bar_windows(len_progresses, title, progress_bar_label):
         root, 
         text="Close", 
         fg = "black", 
+        bg = '#2F4F4F',
         font=("Helvetica", 20),        
         command=lambda: progress_bar_exit(root)
     )
     close_button.grid(column = 1, row = len(len_progresses)+5)
     return root, close_button, progress_bars
 
+def progress_bar_exit(root):
+    root.destroy()
 
-class MyApp_Fitting(threading.Thread):
 
-    def __init__(self, data, threads, close_button, progressbar):
-        self.finished = False
-        self.threads = threads
-        self.data = data
-        self.close_button = close_button
-        # self.progress_label = progress_label
-        self.progressbar = progressbar
-        threading.Thread.__init__(self)
-        self.start()
-
-    def run(self):
-        print(self.data["spec_list"])
-        
-        for fit in self.data["spec_list"]:
-            self.progressbar["value"] += 1
-            # self.progress_label["value"] +=1
-            nff.run_single_fit_function(fit=fit, **self.data)
-        self.finished = True
-        finished = True
-        for thread in self.threads:
-            finished = thread.finished if thread.finished == False else finished
-        if finished:
-            self.close_button.invoke()
-
-class MyApp_Plotting(threading.Thread):
-    def __init__(self, data, threads, close_button, progressbar):
-        self.finished = False
-        self.threads = threads
-        self.data = data
-        self.close_button = close_button
-        self.progressbar = progressbar
-        threading.Thread.__init__(self)
-        self.start()
-
-    def run(self):
-        for r in self.data["speclist"]:
-            single_plot_function(r = r, **self.data)
-            self.progressbar["value"] += 1
-        
-        self.finished = True
-        finished = True
-        for thread in self.threads:
-            finished = thread.finished if thread.finished == False else finished
-        if finished:
-            self.close_button.invoke()
